Remove legacy Kubernetes launcher path
